pyproject.toml
[tool.poetry]
name = "ordo-back"
version = "0.1.0"
description = ""
authors = ["Ordo Development Team"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.5"
djangorestframework = "^3.12.4"
python-dotenv = "^0.19.0"
django-rest-passwordreset = "^1.2.0"
redis = "^3.5.3"
aiohttp = "^3.8.6"
Scrapy = "^2.5.0"
drf-nested-routers = "^0.93.3"
pandas = "^1.3.2"
python-slugify = "^5.0.2"
gunicorn = "^20.1.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
psycopg2-binary = "^2.9.1"
django-ses = "^2.2.1"
django-cors-headers = "^3.8.0"
django-storages = "^1.11.1"
Pillow = "^10.1.0"
django-filter = "^2.4.0"
factory-boy = "^3.2.0"
requests = "^2.26.0"
python-dateutil = "^2.8.2"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.2.0"}
django-credit-cards = "^0.4.1"
django-celery-beat = "^2.2.1"
certifi = "^2023.7.22"
django-nested-admin = "^3.3.3"
django-extensions = "^3.1.3"
aiofiles = "^0.7.0"
djangorestframework-recursive = "^0.1.2"
sentry-sdk = "^1.5.0"
stripe = "^2.63.0"
aioboto3 = "^9.3.1"
celery = "^5.2.6"
selenium = "4.11.2"
webdriver-manager = "^3.8.5"
ebaysdk = "^2.2.0"
Werkzeug = "^2.2.2"
pytest = "^7.2.1"
pytest-django = "^4.5.2"
flashtext = "^2.7"
djangorestframework-simplejwt = "^5.2.2"
coverage = "^7.1.0"
result = "^0.10.0"
pysftp = "^0.2.9"
xmltodict = "^0.13.0"
aws-secretsmanager-caching = "^1.1.1.5"
unicaps = "^1.2.1"
oauthlib = "^3.2.2"
drf-spectacular = "^0.26.2"
django-admin-inline-paginator = "^0.4.0"
regex = "^2023.8.8"
ecs-logging = "^2.1.0"
django-colorfield = "^0.10.1"
slack-sdk = "^3.23.1"
viewstate = "^0.5.3"

[tool.poetry.dev-dependencies]
pre-commit = "^2.21.0"
flake9 = "^3.8.3"

[tool.poetry.scripts]
dbdump = 'scripts.dbdump:main'

[tool.poetry.group.dev.dependencies]
types-factory-boy = "^0.4.1"
pydantic = "^1.10.8"
pytest-env = "^0.8.1"
unittest-parametrize = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 119
extend-ignore = "E203,"
exclude = "*migrations*"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 119

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
